+++ src/constants.ts
@@ -11,7 +11,6 @@ export const projectLanguageSetting: string = 'projectLanguage';
 export const projectLanguageModelSetting: string = 'projectLanguageModel';
 export const funcVersionSetting: string = 'projectRuntime'; // Using this name for the sake of backwards compatability even though it's not the most accurate
 export const projectSubpathSetting: string = 'projectSubpath';
-export const templateFilterSetting: string = 'templateFilter';
 export const deploySubpathSetting: string = 'deploySubpath';
 export const templateVersionSetting: string = 'templateVersion';
 export const preDeployTaskSetting: string = 'preDeployTask';
@@ -22,6 +21,8 @@ export const projectTemplateKeySetting: string = 'projectTemplateKey';
 export const remoteBuildSetting: string = 'scmDoBuildDuringDeployment';
 export const javaBuildTool: string = 'javaBuildTool';
 export const functionSubpathSetting: string = 'functionSubpath';
+export const showBallerinaProjectCreationSetting: string = 'showBallerinaProjectCreation';
+export const mcpProjectTypeSetting: string = 'mcpProjectType';
 
 export const browseItem: IAzureQuickPickItem<undefined> = { label: localize('browse', '$(file-directory) Browse...'), description: '', data: undefined };
 
@@ -36,7 +37,8 @@ export enum ProjectLanguage {
     Python = 'Python',
     TypeScript = 'TypeScript',
     Ballerina = 'Ballerina',
-    Custom = 'Custom'
+    Custom = 'Custom',
+    SelfHostedMCPServer = 'SelfHostedMCPServer'
 }
 
 /**
@@ -73,7 +75,7 @@ export enum JavaBuildTool {
     gradle = 'gradle'
 }
 
-export const ballerinaTomlFileName: string = "Ballerina.toml"
+export const ballerinaTomlFileName: string = "Ballerina.toml";
 export enum BallerinaBackend {
     jvm = 'jvm',
     native = 'native'
@@ -160,6 +162,9 @@ export const extensionVersionKey: string = 'FUNCTIONS_EXTENSION_VERSION';
 export const runFromPackageKey: string = 'WEBSITE_RUN_FROM_PACKAGE';
 export const contentConnectionStringKey: string = 'WEBSITE_CONTENTAZUREFILECONNECTIONSTRING';
 export const contentShareKey: string = 'WEBSITE_CONTENTSHARE';
+export const azureWebJobsFeatureFlags: string = 'AzureWebJobsFeatureFlags';
+export const enableMcpCustomHandlerPreview: string = 'EnableMcpCustomHandlerPreview';
+export const mcpSelfHostedConfigurationProfile: string = 'mcp-custom-handler';
 
 /**
  * The "current" Node.js model is 3 (and assumed, if the number is omitted).
@@ -168,14 +173,14 @@ export const contentShareKey: string = 'WEBSITE_CONTENTSHARE';
  */
 export const nodeDefaultModelVersion: number = 4;
 const nodeDefaultModel: IAzureQuickPickItem<number | undefined> = { data: nodeDefaultModelVersion, label: localize('modelV4', 'Model V4') };
-const nodeV3Model: IAzureQuickPickItem<number | undefined> = { data: undefined, label: localize('modelV3', 'Model V3') }
+const nodeV3Model: IAzureQuickPickItem<number | undefined> = { data: undefined, label: localize('modelV3', 'Model V3') };
 
 export const nodeModels = [nodeDefaultModel, nodeV3Model];
 export const nodeLearnMoreLink = 'https://aka.ms/AzFuncNodeV4';
 
 export const pythonDefaultModelVersion: number = 2;
 const pythonV2Model: IAzureQuickPickItem<number | undefined> = { data: pythonDefaultModelVersion, label: localize('pyModelV2', 'Model V2') };
-const pythonV1Model: IAzureQuickPickItem<number | undefined> = { data: undefined, label: localize('pyModelV1', 'Model V1') }
+const pythonV1Model: IAzureQuickPickItem<number | undefined> = { data: undefined, label: localize('pyModelV1', 'Model V1') };
 
 export const pythonModels = [pythonV2Model, pythonV1Model];
 export const pythonLearnMoreLink = 'https://aka.ms/AAmlyrc';
@@ -219,3 +224,24 @@ export const EventHubsNamespaceResourceType: string = 'namespaces';
 
 export const DurableTaskProvider: string = 'Microsoft.DurableTask';
 export const DurableTaskSchedulersResourceType: string = 'schedulers';
+export type GitHubFileMetadata = {
+    name: string;
+    path: string;
+    sha: string;
+    size: number;
+    url: string;
+    html_url: string;
+    git_url: string;
+    download_url: string;
+    type: string;
+    _links: {
+        self: string;
+        git: string;
+        html: string;
+    };
+};
+
+export enum McpProjectType {
+    McpExtensionServer = 'McpExtensionServer',
+    SelfHostedMcpServer = 'SelfHostedMcpServer'
+}