+++ src/templates/script/getScriptVerifiedTemplateIds.ts
@@ -38,7 +38,9 @@ export function getScriptVerifiedTemplateIds(version: string): (string | RegExp)
             //TODO: Add unit test for EventGridBlobTrigger
             'EventGridBlobTrigger',
             'SqlTrigger',
-            'McpResourceTrigger'
+            'McpToolTrigger',
+                'MCPTrigger',
+                'McpResourceTrigger'
         ]);
 
         // These languages are only supported in v2+ - same functions as JavaScript, with a few minor exceptions that aren't worth distinguishing here